refactor(amazonq): move chat activation into packages/amazonq/
Problem:
Q activation lives in `packages/core/`, so it can't call code defined in
`packages/amazonq/`. This prevents new, Q-only, application code from
being defined in `packages/amazonq/` where it should be.

Solution:
- Move Q activation into `packages/amazonq/`.
- Create an `app/ directory where application logic can live.
