+++ packages/core/src/codewhisperer/index.ts
@@ -92,6 +92,7 @@ export * from './util/zipUtil'
 export * from './util/commonUtil'
 export * from './util/supplementalContext/codeParsingUtil'
 export * from './util/supplementalContext/supplementalContextUtil'
+export * from './util/codewhispererSettings'
 export * as supplementalContextUtil from './util/supplementalContext/supplementalContextUtil'
 export * from './service/diagnosticsProvider'
 export * as diagnosticsProvider from './service/diagnosticsProvider'
