+++ packages/core/src/shared/index.ts
@@ -43,4 +43,5 @@ export * from './utilities/pathUtils'
 export * from './errors'
 export * as messages from './utilities/messages'
 export * as errors from './errors'
+export * as funcUtil from './utilities/functionUtils'
 export { fs } from './fs/fs'
